Added unit tests for user resources

+++ app/users/resources.py
@@ -1,23 +1,11 @@
 import json
 
 from flask import request
-from flask_restful import Resource, reqparse
+from flask_restful import Resource
 
 from app import db, User
-from app.lib.response_util import buildOkResponse
+from app.lib.response_util import buildOkResponse, _returnUser
 from app.lib import status
-
-
-def _returnUser(user):
-    """Private method to convert a user model into a dictionary.
-
-    Args:
-        user - A user model.
-    """
-    return {
-        "id": user.id,
-        "name": user.name
-    }
 
 
 class UserListResource(Resource):
@@ -33,6 +21,10 @@ class UserListResource(Resource):
     def post(self):
         """This method adds a new user and returns the user in an OK response.
         """
+        if not (request.form.get('name') and request.form.get('email') and
+                request.form.get('password')):
+            raise status.BadRequest()
+
         newUser = User(request.form.get('name'),
                        request.form.get('email'),
                        request.form.get('password'))
@@ -54,6 +46,9 @@ class UserResource(Resource):
         """
         user = User.query.get(userId)
 
+        if user is None:
+            raise status.NotFound()
+
         user.name = request.form.get('name') or user.name
         user.password = request.form.get('password') or user.password
 
@@ -67,7 +62,10 @@ class UserResource(Resource):
         Args:
             userId - An integer, primary key that identifies the user.
         """
+
         user = User.query.get(userId)
+        if user is None:
+            raise status.NotFound()
         return buildOkResponse(_returnUser(user))
 
     def delete(self, userId):
@@ -77,7 +75,9 @@ class UserResource(Resource):
             userId - An integer, primary key that identifies the user.
         """
         user = User.query.get(userId)
+        if user is None:
+            raise status.NotFound()
         db.session.delete(user)
-        db.commit()
+        db.session.commit()
 
         return buildOkResponse(None)
